requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
elasticsearch>=8.10.0
confluent-kafka>=2.3.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
yara-python>=4.3.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
prometheus-client>=0.17.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
pytest>=7.4.0
httpx>=0.24.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
fastapi-limiter>=0.1.5
redis>=5.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Cloud Integration Dependencies
boto3>=1.28.0
azure-identity>=1.14.0
azure-monitor-query>=1.2.0
azure-storage-blob>=12.18.0
google-cloud-logging>=3.8.0
google-cloud-storage>=2.13.0
google-cloud-monitoring>=2.15.0

# Security Integration Dependencies
asyncssh>=2.14.0
aiofiles>=23.2.0
aiohttp>=3.9.0
